# openapi.yaml
openapi: 3.0.0
info:
  title: Snake Arena API
  description: API specification for Snake Arena backend, derived from frontend requirements.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Local development server

paths:
  /auth/login:
    post:
      summary: Login user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials

  /auth/signup:
    post:
      summary: Register new user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Email or username already exists

  /auth/logout:
    post:
      summary: Logout user
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logout successful

  /auth/me:
    get:
      summary: Get current user profile
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated

  /leaderboard:
    get:
      summary: Get leaderboard entries
      tags: [Leaderboard]
      parameters:
        - in: query
          name: mode
          schema:
            type: string
            enum: [walls, pass-through]
          description: Filter by game mode
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
          description: Number of entries to return
      responses:
        "200":
          description: List of leaderboard entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"

    post:
      summary: Submit a new score
      tags: [Leaderboard]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitScoreRequest"
      responses:
        "201":
          description: Score submitted successfully
        "401":
          description: Not authenticated

  /games/active:
    get:
      summary: Get active games for spectating
      tags: [Games]
      responses:
        "200":
          description: List of active games
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActiveGame"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
          format: email
      required:
        - id
        - username
        - email

    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
          description: JWT token for authentication
      required:
        - user
        - token

    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password

    SignupRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        username:
          type: string
        password:
          type: string
          minLength: 6
      required:
        - email
        - username
        - password

    LeaderboardEntry:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        score:
          type: integer
        mode:
          type: string
          enum: [walls, pass-through]
        timestamp:
          type: string
          format: date-time
      required:
        - id
        - username
        - score
        - mode
        - timestamp

    SubmitScoreRequest:
      type: object
      properties:
        score:
          type: integer
        mode:
          type: string
          enum: [walls, pass-through]
      required:
        - score
        - mode

    ActiveGame:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        score:
          type: integer
        mode:
          type: string
          enum: [walls, pass-through]
        snake:
          type: array
          items:
            type: object
            properties:
              x:
                type: integer
              y:
                type: integer
        food:
          type: object
          properties:
            x:
              type: integer
            y:
              type: integer
      required:
        - id
        - username
        - score
        - mode
        - snake
        - food